requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
sqlalchemy==2.0.27
pydantic==2.6.3
pydantic-settings==2.2.1
alembic==1.13.1
asyncpg==0.29.0
aiosqlite==0.19.0
greenlet==3.0.3
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
google-generativeai==0.4.0
bcrypt==4.1.2
python-dotenv==1.0.1
email-validator==2.1.1
slowapi==0.1.9
